+++ internal/module/operationq/upload-queue.go
@@ -1,4 +1,4 @@
-package loadq
+package operationq
 
 import (
 	"container/ring"
